requirements.txt
streamlit
pandas
plotly
qrcode
google-generativeai